// lib/api/generated/model/milestoneRead.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Syncnesto API
 * OpenAPI spec version: 0.1.0
 */

/**
 * マイルストーン読み取り時に返すschema。
 */
export interface MilestoneRead {
  title: string;
  description?: string | null;
  target_date: string;
  status?: string;
  id: number;
  project_id: number;
  version: number;
  created_by?: number | null;
  updated_by?: number | null;
  created_at: string;
  updated_at: string;
}